--- a/backend/src/layer/data/database/factory.rs
+++ b/backend/src/layer/data/database/factory.rs
@@ -71,10 +71,7 @@ impl DataAccessFactory {
     /// Returns a object for accessing database requests for api requests.
     #[must_use]
     pub fn get_request_data_access(&self) -> PersistentRequestData {
-        PersistentRequestData {
-            pool: self.pool.clone(),
-            max_weeks_data: self.max_weeks_data,
-        }
+        PersistentRequestData::new(self.pool.clone(), self.max_weeks_data)
     }
 
     /// Returns a object for accessing database requests for authentication.