Until now, for every GraphQL query, there are plenty of database queries (see [n+1 problem](https://async-graphql.github.io/async-graphql/en/dataloader.html)). This PR optimizes this using a data loader which bundles similar queries into a single one.

## ToDo
- [x] queries with fetch streams
- [ ] less data loaders? combining them?

## Benchmarks
### Setting
1000 concurrently started queries requesting the whole meal plan for a day. Run using JMeter. [config.zip](https://github.com/user-attachments/files/18854202/config.zip)

### Results
**Throuhput**: 612 requests / min -> 8197 / min (13x improvement)
**Average query time** 78207 ms -> 4812 ms  (16x improvement)
### Graphics
Before:
<img width="1546" alt="Ergebnisse_alt" src="https://github.com/user-attachments/assets/58d36d16-dfdd-4905-8ac2-94dfe18b680d" />
After:
<img width="1546" alt="Ergebnisse_neu" src="https://github.com/user-attachments/assets/023473cc-e047-47cb-af1c-8290fa0f1594" />

### Less concurrent queries
Single queries are now a bit slower, with 47 ms instead of 27 ms.
At 10 concurrent requests, the advantage is visible: 45 ms instead of 151 ms.
At 100 concurrent requests: 87 ms instead of 1465 ms response time.